+++ pycep/transformer.py
@@ -1004,6 +1004,14 @@ class BicepToJson(Transformer[Token, pycep_typing.BicepJson]):
             },
         }
 
+    def data_uri_to_string(self, args: tuple[pycep_typing.PossibleValue]) -> pycep_typing.DataUriToString:
+        return {
+            "type": "data_uri_to_string",
+            "parameters": {
+                "data_uri_to_convert": args[0],
+            },
+        }
+
     def ends_with(self, args: tuple[pycep_typing.PossibleValue, pycep_typing.PossibleValue]) -> pycep_typing.EndsWith:
         string_to_search, string_to_find = args