+++ tests/test_function/test_parse_string.py
@@ -60,6 +60,19 @@ def test_parse_data_uri() -> None:
     assert_that(result).is_equal_to(expected_result)
 
 
+def test_parse_data_uri_to_string() -> None:
+    # given
+    sub_dir_path = EXAMPLES_DIR / "data_uri_to_string"
+    file_path = sub_dir_path / "main.bicep"
+    expected_result = json.loads((sub_dir_path / "result.json").read_text())
+
+    # when
+    result = BicepParser(file_path=file_path).json()
+
+    # then
+    assert_that(result).is_equal_to(expected_result)
+
+
 def test_parse_ends_with() -> None:
     # given
     sub_dir_path = EXAMPLES_DIR / "ends_with"