+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support `trim` function
 - Support `padLeft` function
 - Support `dataUri` function
+- Support `dataUriToString` function
 
 ## 0.2.0 - 2022-02-13